gatt mtu size change and ram size with ram address
In this project, I need to change the MTU size from 23 to 255, which requires modifying the RAM size and addressing. The information provided by Nordic indicates that this change can be made, so there's no need to worry.

When you need to use it, please specify the pin configuration of the board as I am using a custom nRF52810 board that I created.